requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.3
bcrypt==4.0.1
pyotp==2.8.0
qrcode==7.4
pycryptodome==3.19.0
python-dotenv==1.0.0
